+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Next version
+- Add `nbsp` helper
+- Add `PluckDistinct` model concern
+- Add pagination helpers
 
 ## 1.5.10
 - Cucumber Helpers : add cuprite support